refactor(provider): one event-stream loop for the Docker-API providers
The events reconnect loop existed as two verbatim copies (docker and
dockerswarm) that had already diverged: docker retried forever and never
wrote its Err channel, while swarm gave up after ten attempts and emitted a
terminal error - permanently killing event delivery after roughly one
minute of daemon downtime, even though the daemon coming back is the common
case. A maintainer patching one copy had no way to know the other existed;
the attempt budget was added to swarm only.

Both copies also sent events with a bare channel send. Once the consumer
stops reading (the core watchers nil out the channel on error or return on
cancellation), the producer goroutine blocks on that send forever, leaking
the goroutine and pinning its events HTTP connection. Only the proxmox
provider guarded its sends.

The loop now lives once in pkg/provider/internal/mobyevents with a single
documented policy: reconnect indefinitely with linear backoff capped at
30s, close both channels only on context cancellation, never emit terminal
errors, and guard every send with the context. Docker, Docker Swarm and
(through the docker provider) Podman share it. The
sablier.InstanceEventStream contract is updated to describe the real
semantics.

The old docker unit test moves to the shared package, joined by a leak
regression test (the old loop fails it: the goroutine is still blocked two
seconds after cancellation) and a policy test pinning that the stream
outlives the old ten-attempt budget without a terminal error.
